+++ src/main/java/com/expenses/jonsnow/dto/TransactionSummaryDTO.java
@@ -2,12 +2,10 @@ package com.expenses.jonsnow.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class TransactionSummaryDTO {
-    private BigDecimal lent;
-    private BigDecimal owe;
-    private BigDecimal cashIn;
-    private BigDecimal cashOut;
+    private Long lent;
+    private Long owe;
+    private Long cashIn;
+    private Long cashOut;
 }
